+++ src/Samwise/Runtime/Nodes/IChoosableNode.cs
@@ -1,11 +1,15 @@
 // (c) Copyright 2022 Davide 'PeevishDave' Barbieri
 
+using System.Collections.Generic;
+
 namespace Peevo.Samwise
 {
     public interface IChoosableNode: IDialogueNode
     {
         int OptionsCount { get; }
-        IOption GetOption(int i);
+        Option GetOption(int index);
+
+        IEnumerator<IOption> GetAvailableOptions(IDialogueContext context);
         string CharacterId { get; }
         double? Time { get; }