+++ src/Samwise/Runtime/Nodes/IMultiCaseNode.cs
@@ -1,5 +1,7 @@
 // (c) Copyright 2024 Davide 'PeevishDave' Barbieri
 
+using System.Collections.Generic;
+
 namespace Peevo.Samwise
 {
     public interface IMultiCaseNode